Move buffer sync support into server folder
