+++ extensions/typescript-language-features/src/tsServer/bufferSyncSupport.ts
